+++ BibleAtlas/AppDelegate.swift
@@ -7,6 +7,7 @@
 
 import UIKit
 import CoreData
+import FirebaseCore
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -16,11 +17,11 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
         
-        #if DEBUG
-        setupMockAPI()  // 앱 실행 시 모킹 API 활성화
-        #endif
-        
-        
+//        #if DEBUG
+//        setupMockAPI()  // 앱 실행 시 모킹 API 활성화
+//        #endif
+            
+        FirebaseApp.configure()
         return true
     }
