+++ BibleAtlas/Data/Services/AuthApiService.swift
@@ -12,6 +12,8 @@ import Alamofire
 
 protocol AuthApiServiceProtocol{
     func loginUser(body:AuthPayload) async -> Result<UserResponse,NetworkError>
+    
+    func loginGoogleUser(idToken: String) async -> Result<UserResponse,NetworkError>
 
 }
 
@@ -38,10 +40,25 @@ final public class AuthApiService:AuthApiServiceProtocol {
                 "Authorization": "Basic \(base64)"
             ]
             
-        return await apiClient.postData(url: "\(url)/login", parameters: nil, body: nil, headers: headers)
-            
-        
+            return await apiClient.postData(url: "\(url)/login", parameters: nil, body: nil, headers: headers)
     
+    }
+    
+    func loginGoogleUser(idToken: String) async -> Result<UserResponse, NetworkError> {
+            let body: [String: String] = [
+                "idToken": idToken
+            ]
+            
+            guard let jsonData = try? JSONSerialization.data(withJSONObject: body) else {
+                return .failure(.failToEncode("Invalid JSON"))
+            }
+            
+            return await apiClient.postData(
+                url: "\(url)/google-login",
+                parameters: nil,
+                body: jsonData,
+                headers: ["Content-Type": "application/json"]
+            )
     }
