+++ BibleAtlas/Domain/Protocols/AuthRepositoryProtocol.swift
@@ -9,4 +9,7 @@ import Foundation
 
 protocol AuthRepositoryProtocol {
     func loginUser(body:AuthPayload) async -> Result<UserResponse,NetworkError>
+    
+    func loginGoogleUser(idToken: String) async -> Result<UserResponse,NetworkError>
+    
 }
