+++ BibleAtlas/GoogleService-Info.plist
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>API_KEY</key>
+	<string>AIzaSyCZnDchhSKTk5jAkz4F7eBWkZ1kH5itlzo</string>
+	<key>GCM_SENDER_ID</key>
+	<string>808096669363</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>com.seongyeon.BibleAtlas</string>
+	<key>PROJECT_ID</key>
+	<string>bible-atlas</string>
+	<key>STORAGE_BUCKET</key>
+	<string>bible-atlas.firebasestorage.app</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:808096669363:ios:9c95ef6c2d4c69b9fedac6</string>
+</dict>
+</plist>
