+++ BibleAtlas/Presentation/ViewControllers/HomeContentViewController.swift
@@ -1,0 +1,31 @@
+//
+//  HomeContentViewController.swift
+//  BibleAtlas
+//
+//  Created by 배성연 on 6/21/25.
+//
+
+import UIKit
+import RxSwift
+import RxCocoa
+
+class HomeContentViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
