+++ BibleAtlas/Presentation/ViewModels/HomeBottomSheetVM/SearchReadyViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  SearchReadyViewModel.swift
+//  BibleAtlas
+//
+//  Created by 배성연 on 6/21/25.
+//
+
+import Foundation
+
+protocol SearchReadyViewModelProtocol {}
+
+final class SearchReadyViewModel: SearchReadyViewModelProtocol {
+    // 추후 검색 준비 상태 관련 로직이 들어올 수 있음
+}
