+++ BibleAtlas/Presentation/ViewModels/HomeBottomSheetVM/SearchResultViewModel.swift
@@ -1,26 +1,25 @@
 //
-//  SearchViewModel.swift
+//  SearchResultViewModel.swift
 //  BibleAtlas
 //
-//  Created by 배성연 on 6/17/25.
+//  Created by 배성연 on 6/21/25.
 //
+
+import Foundation
 
 import Foundation
 import RxSwift
 import RxRelay
 import RxCocoa
 
-protocol SearchBottomSheetViewModelProtocol {
-    func transform(input:SearchBottomSheetViewModel.Input) -> SearchBottomSheetViewModel.Output
+protocol SearchResultViewModelProtocol {
+    func transform(input:SearchResultViewModel.Input) -> SearchResultViewModel.Output
 }
 
 
-
-final class SearchBottomSheetViewModel:SearchBottomSheetViewModelProtocol {
+final class SearchResultViewModel:SearchResultViewModelProtocol {
     private let disposeBag = DisposeBag();
-            
-    private let screenMode$ = BehaviorRelay<HomeScreenMode>(value: .home)
-    
+
     private weak var navigator: BottomSheetNavigator?
 
     private let placeUsecase:PlaceUsecaseProtocol?
@@ -33,99 +32,70 @@ final class SearchBottomSheetViewModel:SearchBottomSheetViewModelProtocol {
     private let isSearching$ = BehaviorRelay<Bool>(value: false)
     private let isFetchingNext$ = BehaviorRelay<Bool>(value: false);
 
-        
-    private let isSearchingMode$ = BehaviorRelay<Bool>(value: false);
-    
-    private let keyword$ = BehaviorRelay<String>(value: "");
-    
+    private let isSearchingMode$:Observable<Bool>
+    private let keyword$:Observable<String>
+    private let cancelButtonTapped$:Observable<Void>
     
     
-    init(navigator: BottomSheetNavigator? = nil, placeUsecase: PlaceUsecaseProtocol?) {
+    
+    init(navigator: BottomSheetNavigator? = nil, placeUsecase: PlaceUsecaseProtocol?, isSearchingMode$:Observable<Bool>, keyword$:Observable<String>, cancelButtonTapped$:Observable<Void> ) {
         self.navigator = navigator
         self.placeUsecase = placeUsecase
+        self.isSearchingMode$ = isSearchingMode$
+        self.keyword$ = keyword$
+        self.cancelButtonTapped$ = cancelButtonTapped$
+        
     }
         
     
     func transform(input:Input) -> Output {
-            
-    
-        keyword$
+
+        let debouncedKeyword$ = keyword$
             .map { $0.trimmingCharacters(in: .whitespacesAndNewlines) }
             .debounce(.milliseconds(300), scheduler: MainScheduler.instance)
             .distinctUntilChanged()
-            .subscribe(onNext: {[weak self] keyword in
+        
+        Observable.combineLatest(debouncedKeyword$, isSearchingMode$)
+            .subscribe(onNext: {[weak self] keyword, isSearchingMode in
                 guard let self = self else {
                     return;
                 }
-                
-                if(!self.isSearchingMode$.value){
+                if(!isSearchingMode){
                     return
                 }
                 
-
                 
                 if keyword.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
                     self.places$.accept([])
                     self.error$.accept(nil)
                     self.isSearching$.accept(false)
-                    self.screenMode$.accept(.searchReady)
                     return
                 }
                 
                 self.pagination.reset()
                 self.getPlaces(keyword: keyword)
-                self.screenMode$.accept(.searching)
-
                 
             })
             .disposed(by: disposeBag)
         
-        isSearchingMode$.asObservable().subscribe(onNext: {[weak self] isSearchingMode in
-            guard let self = self else { return }
-            if(isSearchingMode){
-                if(self.keyword$.value.isEmpty){
-                    self.screenMode$.accept(.searchReady)
-                }
-                else{
-                    self.screenMode$.accept(.searching)
-                }
-            }
-            else{
-                self.screenMode$.accept(.home)
-                self.keyword$.accept("")
-            }
-        }).disposed(by: disposeBag)
-        
-        input.cancelButtonTapped$
-            .subscribe(onNext: {[weak self] in
-                self?.places$.accept([])
-                self?.error$.accept(nil)
-                self?.pagination.reset();
-                    
-                self?.isSearchingMode$.accept(false)
-                
-            })
-            .disposed(by: disposeBag)
-        
-        input.editingDidBegin$
-            .subscribe(onNext: { [weak self] in
-                self?.isSearchingMode$.accept(true);
-            })
-            .disposed(by: disposeBag)
-        
-        
-        input.bottomReached$
+       
+            
+        let debouncedBottomReached$ = input.bottomReached$
             .debounce(.microseconds(500), scheduler: MainScheduler.instance)
-            .subscribe(onNext:{ [weak self] in
+        
+        Observable.combineLatest(debouncedBottomReached$, keyword$)
+            .subscribe(onNext: { [weak self] _, keyword in
                 guard let self = self else { return }
-                if(self.keyword$.value.isEmpty){
+                if(keyword.isEmpty){
                     return;
                 }
                 
-                self.getMorePlaces(keyword: self.keyword$.value)
-                
+                self.getMorePlaces(keyword: keyword)
+
+
             })
             .disposed(by: disposeBag)
+        
         
         input.placeCellSelected$
             .subscribe(onNext:{ [weak self] placeId in
@@ -136,7 +106,7 @@ final class SearchBottomSheetViewModel:SearchBottomSheetViewModelProtocol {
             })
             .disposed(by: disposeBag)
         
-        return Output(places$: places$.asObservable(), error$: error$.asObservable(), isSearching$: isSearching$.asObservable(), isFetchingNext$: isFetchingNext$.asObservable(),screenMode$: screenMode$.asObservable(), keywordRelay$: keyword$, keywordText$: keyword$.asDriver(onErrorJustReturn: ""), isSearchingMode$: isSearchingMode$.asObservable())
+        return Output(places$: places$.asObservable(), error$: error$.asObservable(), isSearching$: isSearching$.asObservable(), isFetchingNext$: isFetchingNext$.asObservable(), isSearchingMode$: isSearchingMode$.asObservable())
     }
     
     
@@ -199,12 +169,9 @@ final class SearchBottomSheetViewModel:SearchBottomSheetViewModelProtocol {
     
     
     public struct Input{
-        let cancelButtonTapped$:Observable<Void>
-        let editingDidBegin$:Observable<Void>
 //        let refetchButtonTapped$:Observable<Void>
         let bottomReached$:Observable<Void>
         let placeCellSelected$:Observable<String>
-        
     }
     
     
@@ -213,9 +180,6 @@ final class SearchBottomSheetViewModel:SearchBottomSheetViewModelProtocol {
         let error$:Observable<NetworkError?>
         let isSearching$:Observable<Bool>
         let isFetchingNext$:Observable<Bool>
-        let screenMode$:Observable<HomeScreenMode>
-        let keywordRelay$: BehaviorRelay<String>
-        let keywordText$: Driver<String>
         let isSearchingMode$:Observable<Bool>
     }
