+++ BibleAtlas/SceneDelegate.swift
@@ -6,8 +6,8 @@
 //
 
 import UIKit
-import UBottomSheet
-import PanModal
+import GoogleSignIn
+
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
     var window: UIWindow?
@@ -90,6 +90,14 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         
        
     }
+    
+    func scene(_ scene: UIScene,
+               openURLContexts URLContexts: Set<UIOpenURLContext>) {
+
+        guard let url = URLContexts.first?.url else { return }
+        GIDSignIn.sharedInstance.handle(url)
+    }
+    
 
     func sceneDidDisconnect(_ scene: UIScene) {
     }
